# Implement the new shared memory interface