--- a/src/infer_request.cc
+++ b/src/infer_request.cc
@@ -1,4 +1,4 @@
-// Copyright 2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2021-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -25,11 +25,11 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "infer_request.h"
-#include <boost/interprocess/sync/scoped_lock.hpp>
 
+#include <boost/interprocess/sync/scoped_lock.hpp>
 #include "pb_utils.h"
+#include "scoped_defer.h"
 #ifdef TRITON_PB_STUB
-#include "infer_response.h"
 #include "pb_stub.h"
 #endif
 
@@ -95,77 +95,187 @@ InferRequest::SetFlags(uint32_t flags)
   flags_ = flags;
 }
 
-void
-InferRequest::SaveToSharedMemory(
-    std::unique_ptr<SharedMemory>& shm_pool, Request* request_shm)
+bi::managed_external_buffer::handle_t
+InferRequest::ShmHandle()
 {
-  request_shm->correlation_id = this->CorrelationId();
-  off_t id_offset;
-  SaveStringToSharedMemory(shm_pool, id_offset, this->RequestId().c_str());
-  request_shm->id = id_offset;
-  request_shm->requested_output_count = this->RequestedOutputNames().size();
-  off_t requested_output_names_offset;
-  off_t* requested_output_names;
-  shm_pool->Map(
-      (char**)&requested_output_names,
-      sizeof(off_t) * request_shm->requested_output_count,
-      requested_output_names_offset);
-  request_shm->flags = Flags();
+  return shm_handle_;
+}
 
-  request_shm->requested_output_names = requested_output_names_offset;
+void
+InferRequest::SaveToSharedMemory(std::unique_ptr<SharedMemoryManager>& shm_pool)
+{
+  AllocatedSharedMemory<char> infer_request_shm = shm_pool->Construct<char>(
+      sizeof(InferRequestShm) +
+      (RequestedOutputNames().size() *
+       sizeof(bi::managed_external_buffer::handle_t)) +
+      (Inputs().size() * sizeof(bi::managed_external_buffer::handle_t)) +
+      PbString::ShmStructSize(ModelName()) +
+      PbString::ShmStructSize(RequestId()));
+
+  infer_request_shm_ptr_ =
+      reinterpret_cast<InferRequestShm*>(infer_request_shm.data_.get());
+  infer_request_shm_ptr_->correlation_id = CorrelationId();
+  infer_request_shm_ptr_->input_count = Inputs().size();
+  infer_request_shm_ptr_->model_version = model_version_;
+  infer_request_shm_ptr_->requested_output_count =
+      RequestedOutputNames().size();
+  infer_request_shm_ptr_->flags = Flags();
+
+  output_names_handle_shm_ptr_ =
+      reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+          reinterpret_cast<char*>(infer_request_shm_ptr_) +
+          sizeof(InferRequestShm));
+
+  // [FIXME] This could also be a part of the single allocated memory for this
+  // object.
   size_t i = 0;
+  std::vector<std::unique_ptr<PbString>> requested_output_names_shm;
   for (auto& requested_output_name : requested_output_names_) {
-    SaveStringToSharedMemory(
-        shm_pool, requested_output_names[i], requested_output_name.c_str());
+    std::unique_ptr<PbString> requested_output_name_shm =
+        PbString::Create(shm_pool, requested_output_name);
+    output_names_handle_shm_ptr_[i] = requested_output_name_shm->ShmHandle();
+    requested_output_names_shm.emplace_back(
+        std::move(requested_output_name_shm));
     i++;
   }
 
-  request_shm->requested_input_count = this->Inputs().size();
-  request_shm->model_version = this->model_version_;
-  SaveStringToSharedMemory(
-      shm_pool, request_shm->model_name, this->model_name_.c_str());
+  input_tensors_handle_ptr_ =
+      reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+          reinterpret_cast<char*>(output_names_handle_shm_ptr_) +
+          sizeof(bi::managed_external_buffer::handle_t) *
+              RequestedOutputNames().size());
+  i = 0;
+  for (auto& input : Inputs()) {
+    input_tensors_handle_ptr_[i] = input->ShmHandle();
+    i++;
+  }
+
+  size_t model_name_offset =
+      sizeof(InferRequestShm) +
+      (RequestedOutputNames().size() *
+       sizeof(bi::managed_external_buffer::handle_t)) +
+      (Inputs().size() * sizeof(bi::managed_external_buffer::handle_t));
+
+  std::unique_ptr<PbString> model_name_shm = PbString::Create(
+      ModelName(),
+      reinterpret_cast<char*>(infer_request_shm_ptr_) + model_name_offset,
+      infer_request_shm.handle_ + model_name_offset);
+
+  size_t request_id_offset =
+      model_name_offset + PbString::ShmStructSize(ModelName());
+  std::unique_ptr<PbString> request_id_shm = PbString::Create(
+      RequestId(),
+      reinterpret_cast<char*>(infer_request_shm_ptr_) + request_id_offset,
+      infer_request_shm.handle_ + request_id_offset);
+
+  // Save the references to shared memory.
+  infer_request_shm_ = std::move(infer_request_shm);
+  request_id_shm_ = std::move(request_id_shm);
+  model_name_shm_ = std::move(model_name_shm);
+  shm_handle_ = infer_request_shm_.handle_;
+  requested_output_names_shm_ = std::move(requested_output_names_shm);
 }
 
 std::unique_ptr<InferRequest>
 InferRequest::LoadFromSharedMemory(
-    std::unique_ptr<SharedMemory>& shm_pool, off_t request_offset,
-    std::shared_ptr<std::mutex>& cuda_ipc_open_mutex,
-    std::shared_ptr<std::mutex>& cuda_ipc_close_mutex)
+    std::unique_ptr<SharedMemoryManager>& shm_pool,
+    bi::managed_external_buffer::handle_t request_handle, bool open_cuda_handle)
 {
-  Request* request;
-  shm_pool->MapOffset((char**)&request, request_offset);
+  AllocatedSharedMemory<char> infer_request_shm =
+      shm_pool->Load<char>(request_handle);
+  InferRequestShm* infer_request_shm_ptr =
+      reinterpret_cast<InferRequestShm*>(infer_request_shm.data_.get());
 
-  char* id = nullptr;
-  LoadStringFromSharedMemory(shm_pool, request->id, id);
+  std::vector<std::unique_ptr<PbString>> requested_output_names_shm;
+  uint32_t requested_output_count =
+      infer_request_shm_ptr->requested_output_count;
 
-  uint32_t requested_input_count = request->requested_input_count;
-
-  std::vector<std::shared_ptr<PbTensor>> py_input_tensors;
-  for (size_t input_idx = 0; input_idx < requested_input_count; ++input_idx) {
-    std::shared_ptr<PbTensor> pb_input_tensor = PbTensor::LoadFromSharedMemory(
-        shm_pool, request->inputs + sizeof(Tensor) * input_idx,
-        cuda_ipc_open_mutex, cuda_ipc_close_mutex);
-    py_input_tensors.emplace_back(std::move(pb_input_tensor));
-  }
-
-  std::vector<std::string> requested_output_names;
-  uint32_t requested_output_count = request->requested_output_count;
-  off_t* output_names;
-  shm_pool->MapOffset((char**)&output_names, request->requested_output_names);
+  bi::managed_external_buffer::handle_t* output_names_handle_shm_ptr =
+      reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+          (reinterpret_cast<char*>(infer_request_shm_ptr) +
+           sizeof(InferRequestShm)));
 
   for (size_t output_idx = 0; output_idx < requested_output_count;
        ++output_idx) {
-    char* output_name = nullptr;
-    LoadStringFromSharedMemory(shm_pool, output_names[output_idx], output_name);
-    requested_output_names.emplace_back(output_name);
+    std::unique_ptr<PbString> pb_string = PbString::LoadFromSharedMemory(
+        shm_pool, output_names_handle_shm_ptr[output_idx]);
+    requested_output_names_shm.emplace_back(std::move(pb_string));
   }
 
-  char* model_name;
-  LoadStringFromSharedMemory(shm_pool, request->model_name, model_name);
-  return std::make_unique<InferRequest>(
-      id, request->correlation_id, std::move(py_input_tensors),
-      requested_output_names, model_name, request->model_version,
-      request->flags);
+  bi::managed_external_buffer::handle_t* input_names_handle_shm_ptr =
+      reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+          (reinterpret_cast<char*>(infer_request_shm_ptr) +
+           sizeof(InferRequestShm) +
+           (infer_request_shm_ptr->requested_output_count *
+            sizeof(bi::managed_external_buffer::handle_t))));
+
+  std::vector<std::shared_ptr<PbTensor>> input_tensors;
+  for (size_t input_idx = 0; input_idx < infer_request_shm_ptr->input_count;
+       ++input_idx) {
+    std::shared_ptr<PbTensor> input_tensor = PbTensor::LoadFromSharedMemory(
+        shm_pool, input_names_handle_shm_ptr[input_idx], open_cuda_handle);
+    input_tensors.emplace_back(std::move(input_tensor));
+  }
+
+  size_t model_name_offset =
+      sizeof(InferRequestShm) +
+      (requested_output_count * sizeof(bi::managed_external_buffer::handle_t)) +
+      (infer_request_shm_ptr->input_count *
+       sizeof(bi::managed_external_buffer::handle_t));
+
+  std::unique_ptr<PbString> model_name_shm = PbString::LoadFromSharedMemory(
+      request_handle + model_name_offset,
+      reinterpret_cast<char*>(infer_request_shm_ptr) + model_name_offset);
+
+  size_t request_id_offset = model_name_offset + model_name_shm->Size();
+  std::unique_ptr<PbString> request_id_shm = PbString::LoadFromSharedMemory(
+      request_handle + request_id_offset,
+      reinterpret_cast<char*>(infer_request_shm_ptr) + request_id_offset);
+
+  return std::unique_ptr<InferRequest>(new InferRequest(
+      infer_request_shm, request_id_shm, requested_output_names_shm,
+      model_name_shm, input_tensors));
+}
+
+InferRequest::InferRequest(
+    AllocatedSharedMemory<char>& infer_request_shm,
+    std::unique_ptr<PbString>& request_id_shm,
+    std::vector<std::unique_ptr<PbString>>& requested_output_names_shm,
+    std::unique_ptr<PbString>& model_name_shm,
+    std::vector<std::shared_ptr<PbTensor>>& input_tensors)
+    : infer_request_shm_(std::move(infer_request_shm)),
+      request_id_shm_(std::move(request_id_shm)),
+      requested_output_names_shm_(std::move(requested_output_names_shm)),
+      model_name_shm_(std::move(model_name_shm))
+{
+  infer_request_shm_ptr_ =
+      reinterpret_cast<InferRequestShm*>(infer_request_shm_.data_.get());
+  output_names_handle_shm_ptr_ =
+      reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+          reinterpret_cast<char*>(infer_request_shm_ptr_) +
+          sizeof(InferRequestShm));
+  input_tensors_handle_ptr_ =
+      reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+          reinterpret_cast<char*>(infer_request_shm_ptr_) +
+          sizeof(InferRequestShm) +
+          sizeof(bi::managed_external_buffer::handle_t) *
+              infer_request_shm_ptr_->requested_output_count);
+  inputs_ = std::move(input_tensors);
+
+  std::vector<std::string> requested_output_names;
+  for (size_t output_idx = 0;
+       output_idx < infer_request_shm_ptr_->requested_output_count;
+       ++output_idx) {
+    auto& pb_string = requested_output_names_shm_[output_idx];
+    requested_output_names.emplace_back(pb_string->String());
+  }
+
+  request_id_ = request_id_shm_->String();
+  requested_output_names_ = std::move(requested_output_names);
+  model_name_ = model_name_shm_->String();
+  flags_ = infer_request_shm_ptr_->flags;
+  model_version_ = infer_request_shm_ptr_->model_version;
+  correlation_id_ = infer_request_shm_ptr_->correlation_id;
 }
 
 #ifdef TRITON_PB_STUB
@@ -175,44 +285,57 @@ InferRequest::Exec()
   ResponseBatch* response_batch = nullptr;
   bool responses_is_set = false;
   std::unique_ptr<Stub>& stub = Stub::GetOrCreateInstance();
-  std::unique_ptr<SharedMemory>& shm_pool = stub->GetSharedMemory();
+  std::unique_ptr<SharedMemoryManager>& shm_pool = stub->SharedMemory();
+  bi::managed_external_buffer::handle_t* response_handle = nullptr;
+
+  PythonBackendException pb_exception(std::string{});
+  std::unique_ptr<IPCMessage> ipc_message;
+
+  AllocatedSharedMemory<char> request_batch;
+  ScopedDefer data_load_complete(std::bind([&ipc_message] {
+    bi::scoped_lock<bi::interprocess_mutex> lock{
+        *(ipc_message->ResponseMutex())};
+    ipc_message->ResponseCondition()->notify_all();
+  }));
 
   try {
     py::gil_scoped_release release;
-    std::unique_ptr<IPCMessage> ipc_message =
-        std::make_unique<IPCMessage>(shm_pool, true /* inline_response */);
+    ipc_message = IPCMessage::Create(shm_pool, true /* inline_response */);
     bool has_exception = false;
     PythonBackendException pb_exception(std::string{});
 
     ipc_message->Command() =
         PYTHONSTUB_CommandType::PYTHONSTUB_InferExecRequest;
 
-    RequestBatch* request_batch;
-    shm_pool->Map(
-        (char**)&request_batch, sizeof(RequestBatch), ipc_message->Args());
-    request_batch->batch_size = 1;
+    request_batch = shm_pool->Construct<char>(
+        sizeof(RequestBatch) + sizeof(bi::managed_external_buffer::handle_t));
 
-    Request* request;
-    shm_pool->Map((char**)&request, sizeof(Request), request_batch->requests);
+    RequestBatch* request_batch_shm_ptr =
+        reinterpret_cast<RequestBatch*>(request_batch.data_.get());
+    request_batch_shm_ptr->batch_size = 1;
+    ipc_message->Args() = request_batch.handle_;
 
-    request->requested_input_count = this->Inputs().size();
-    Tensor* tensors;
+    bi::managed_external_buffer::handle_t* requests_shm =
+        reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+            request_batch.data_.get() + sizeof(RequestBatch));
+    request_batch_shm_ptr->batch_size = 1;
+
     bool has_gpu_tensor = false;
-    shm_pool->Map(
-        (char**)&tensors, sizeof(Tensor) * request->requested_input_count,
-        request->inputs);
-
     size_t i = 0;
     for (auto& input_tensor : inputs_) {
-      input_tensor->SaveToSharedMemory(
-          shm_pool, &tensors[i], true /* copy_cpu */, false /* copy_gpu */);
+      input_tensor->SaveToSharedMemory(shm_pool, false /* copy_gpu */);
       if (!input_tensor->IsCPU()) {
         has_gpu_tensor = true;
       }
       ++i;
     }
 
-    SaveToSharedMemory(shm_pool, request);
+    SaveToSharedMemory(shm_pool);
+
+    // Save the shared memory offset of the request.
+    *requests_shm = ShmHandle();
+
+    // Send the BLS request to the parent process and wait for the response.
     {
       bi::scoped_lock<bi::interprocess_mutex> lock{
           *(ipc_message->ResponseMutex())};
@@ -220,21 +343,31 @@ InferRequest::Exec()
       ipc_message->ResponseCondition()->wait(lock);
     }
 
+    // Additional round trip required for asking the stub process
+    // to fill in the GPU tensor buffers
     if (has_gpu_tensor) {
+      AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+          gpu_buffers_handle =
+              shm_pool->Load<bi::managed_external_buffer::handle_t>(
+                  request_batch_shm_ptr->gpu_buffers_handle);
       try {
+#ifdef TRITON_ENABLE_GPU
+        size_t i = 0;
         for (auto& input_tensor : this->Inputs()) {
           if (!input_tensor->IsCPU()) {
-#ifdef TRITON_ENABLE_GPU
-            input_tensor->SetCudaIpcMutexes(
-                stub->CudaIpcOpenMutex(), stub->CudaIpcCloseMutex());
-            input_tensor->LoadGPUData(shm_pool);
-#endif  // TRITON_ENABLE_GPU
+            std::unique_ptr<PbMemory> dst_buffer =
+                PbMemory::LoadFromSharedMemory(
+                    shm_pool, (gpu_buffers_handle.data_.get())[i],
+                    true /* open cuda handle */);
+            PbMemory::CopyBuffer(dst_buffer, input_tensor->Memory());
+            ++i;
           }
         }
+#endif  // TRITON_ENABLE_GPU
       }
       catch (const PythonBackendException& exception) {
         // We need to catch the exception here. Otherwise, we will not notify
-        // the main process and it will wait for the resposne forever.
+        // the main process and it will wait for the response forever.
         pb_exception = exception;
         has_exception = true;
       }
@@ -255,17 +388,23 @@ InferRequest::Exec()
 
     // Get the response for the current message.
     std::unique_ptr<IPCMessage> bls_response = IPCMessage::LoadFromSharedMemory(
-        shm_pool, ipc_message->RequestOffset());
-    shm_pool->MapOffset((char**)&response_batch, bls_response->Args());
-    responses_is_set = true;
+        shm_pool, ipc_message->ResponseHandle());
 
+    AllocatedSharedMemory<char> response_batch_shm =
+        shm_pool->Load<char>(bls_response->Args());
+    response_batch =
+        reinterpret_cast<ResponseBatch*>(response_batch_shm.data_.get());
+    response_handle = reinterpret_cast<bi::managed_external_buffer::handle_t*>(
+        response_batch_shm.data_.get() + sizeof(ResponseBatch));
+
+    responses_is_set = true;
     if (response_batch->has_error) {
       if (response_batch->is_error_set) {
-        char* err_string;
-        LoadStringFromSharedMemory(shm_pool, response_batch->error, err_string);
+        std::unique_ptr<PbString> pb_string =
+            PbString::LoadFromSharedMemory(shm_pool, response_batch->error);
         return std::make_unique<InferResponse>(
             std::vector<std::shared_ptr<PbTensor>>{},
-            std::make_shared<PbError>(err_string));
+            std::make_shared<PbError>(pb_string->String()));
       } else {
         return std::make_unique<InferResponse>(
             std::vector<std::shared_ptr<PbTensor>>{},
@@ -283,8 +422,7 @@ InferRequest::Exec()
   if (responses_is_set) {
     std::unique_ptr<InferResponse> infer_response =
         InferResponse::LoadFromSharedMemory(
-            shm_pool, response_batch->responses, stub->CudaIpcOpenMutex(),
-            stub->CudaIpcCloseMutex());
+            shm_pool, *response_handle, true /* open cuda handle */);
 
     return infer_response;
   } else {
@@ -296,4 +434,5 @@ InferRequest::Exec()
 }
 
 #endif
+
 }}}  // namespace triton::backend::python