Performance results before and after these changes for BLS:

|  Scenario  |  Latency  (before) |  Latency (after)  | Throughput (before) |  Throughput (after)  |
|:--- |:--- | :--- |:--- | :--- |
| 4mb_payload_latency | 11997 | 10617 |  83.4 | 94.2 |
| 1kb_payload_latency | 520 |  533 |  1918.6 | 1871.6 |
| 4b_payload_latency |  503 |  506 |  1981.6 | 1969.6  |
| 4mb_payload_throughput | 471614 | 460934 |  136.4 | 138.8 |
| 1kb_payload_throughput | 18078 | 18098 |  3539 |  3535.2 |
| 4b_payload_throughput |  17348 |  17399 |  3689 | 3677.2 |

The *_latency scenarios use a single instance count with the concurrency of 1. The *_throughput scenarios use 2 instance counts with concurrency of 64. The numbers were collected using the Perf Analyzer C-API. The results are almost the same before and after these changes. The only curious result is the 4mb_payload_latency. I applied one performance improvement as a part of this change which was to not rewrite the tensor to shared memory if it is already saved to shared memory. I think this is why we are seeing slight improvement in the results.

This PR is just the combination of all the previous changes and contains no new commit.